+++ NEWS.md
@@ -1,11 +1,11 @@
 # CFAEpiNow2Pipeline (development version)
 
-
+* Added function families to documentation
 * Renamed file containing diagnostic functions
-
 * Change formatting of metadata values to be atomic.
 * Add `blob_storage_container` as a field to the metadata.
 * Use empty string for paths when non-existant.
+* Add function families
 * Populated the default values of the metadata to be saved.
 * Creating a Config class to make syncing configuration differences easier.
 * Add a JSON reader for the Config class.